functional balances UI

+++ app/src/main/java/com/example/plantastic/models/Groups.kt
@@ -11,5 +11,5 @@ data class Groups(
     val name: String? = null,
     val admins: ArrayList<String>? = null,
     val latestMessage: Message? = null,
-    var balances:HashMap<String, HashMap<String, Int>>? = null
+    var balances:HashMap<String, HashMap<String, Double>>? = null
 )
